feat: token 발급 훅 분리
